+++ HISTORY.rst
@@ -8,6 +8,7 @@ History
 
 * Block as expected on start, even if already started.
 * Raise runtime error if starting the reader without a message handler.
+* Add on_close signal to the reader.
 
 0.1.2 (2014-07-08)
 ~~~~~~~~~~~~~~~~~~